# docs/swagger.yaml
basePath: /
definitions:
...
  model.Categoria:
    properties:
      atualizado_em:
        type: string
      criado_em:
        type: string
      id:
        type: string
      nome:
        type: string
      status:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestão de chamados
  title: Gestor de Chamados API
  version: "1.0"
paths:
  /categorias/ativar/{id}:
    patch:
      consumes:
      - application/json
      description: Ativa uma categoria pelo ID.
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Categoria'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Ativar categoria
      tags:
      - Categorias
  /categorias/atualizar/{id}:
    put:
      consumes:
      - application/json
      description: Atualiza uma categoria existente pelo ID com os dados fornecidos
        no corpo da requisição.
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      - description: Categoria
        in: body
        name: categoria
        required: true
        schema:
          $ref: '#/definitions/model.Categoria'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Atualizar categoria
      tags:
      - Categorias
  /categorias/buscar-por-id/{id}:
    get:
      consumes:
      - application/json
      description: Retorna uma categoria pelo ID
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Categoria'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Buscar categoria por ID
      tags:
      - Categorias
  /categorias/buscar-por-nome/{nome}:
    get:
      consumes:
      - application/json
      description: Retorna uma categoria pelo nome
      parameters:
      - description: Nome da categoria
        in: path
        name: nome
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Categoria'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Buscar categoria por nome
      tags:
      - Categorias
  /categorias/buscar-tudo:
    get:
      consumes:
      - application/json
      description: Retorna lista paginada de categorias
      parameters:
      - description: Página
        in: query
        name: pagina
        type: integer
      - description: Limite
        in: query
        name: limite
        type: integer
      - description: Busca
        in: query
        name: busca
        type: string
      - description: Status
        in: query
        name: status
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Categoria'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Listar todas as categorias
      tags:
      - Categorias
  /categorias/criar:
    post:
      consumes:
      - application/json
      description: Cria uma nova categoria com dados fornecidos no corpo da requisição.
      parameters:
      - description: Categoria
        in: body
        name: categoria
        required: true
        schema:
          $ref: '#/definitions/model.Categoria'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Categoria'
        "400":
          description: Bad Request
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Criar uma nova categoria
      tags:
      - Categorias
  /categorias/desativar/{id}:
    delete:
      consumes:
      - application/json
      description: Desativa (soft delete) uma categoria pelo ID.
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Categoria'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Desativar categoria
      tags:
      - Categorias
  /categorias/listar-completa:
    get:
      consumes:
      - application/json
      description: Retorna lista completa de categorias sem paginação
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Categoria'
            type: array
        "405":
          description: Method Not Allowed
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Listar todas as categorias (completa)
      tags:
      - Categorias
...
swagger: "2.0"